+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 
 * Add `ruby-3.0` to CI
 * Add `codecov` branch coverage check in CI
+* Add new methods:
+  * `OnlyofficeLoggerHelper.green_log()`
+  * `OnlyofficeLoggerHelper.red_log()`
 
 ## 1.1.0 (2020-11-17)